requirements.txt
PyYAML
requests
pandas
python-slugify
duckdb
beautifulsoup4
fastapi
uvicorn[standard]
fastapi
uvicorn[standard]
